Post update in home page

+++ application/controllers/Home.php
@@ -58,10 +58,13 @@ class Home extends CI_Controller {
                 if(!empty($_FILES['media_post']['name'])) {
                     $config['upload_path'] = $this->config->item('post_media_path');
                     
-                    $config['allowed_types'] = 'jpeg|gif|jpg|png|mp4|ogg|webm|mp3';
+                    $config['allowed_types'] = '*';
+                    
+                    /* $config['allowed_types'] = 'jpeg|gif|jpg|png|mp4|ogg|webm|mp3';
                     if(in_array($_SERVER['REMOTE_ADDR'], ['127.0.0.1', "::1"])){
                         $config['allowed_types'] = 'flv|FLV|mkv|MKV|jpeg|gif|jpg|png|mp4|ogg|ogv|avi|mp3';
-                    }
+                    }*/
+                    
                     $config['max_size']	= '1024480';
                     $config['max_width']  = 0;
                     $config['max_height']  = 0;
@@ -82,25 +85,26 @@ class Home extends CI_Controller {
                     { 
                         $mediaArray = $this->upload->data();
                         $media_file = $mediaArray['file_name'];
-                        $ext = pathinfo($media_file, PATHINFO_EXTENSION);
                         
-                         $media_type = 3; // for voice
-                         
-                        if($mediaArray['is_image'] != 1 && in_array($ext, ['flv','FLV','mkv','MKV','mp4','ogg','ogv','avi'])) {
+                         $file_type = explode('/', $mediaArray['file_type']);
+                        if($file_type[0] == 'video') {
                             $media_type = 2; // for video
-                            if(in_array($_SERVER['REMOTE_ADDR'], ['127.0.0.1', "::1"])){
-                                $conv_file_name = 'con_'.time().'.mp4';
-                                $cont_file_path = $this->config->item('post_media_path').$conv_file_name;
-                                shell_exec("ffmpeg -i ".$this->config->item('post_media_path').$media_file." -f mp4 -s 500x400 -strict -2  ".$cont_file_path." 2>&1");
-                                // for deleting the source file
-                                if(file_exists($this->config->item('post_media_path').$media_file)) {
-                                    unlink($this->config->item('post_media_path').$media_file);
-                                }
-                                $media_file = $conv_file_name;
-                            }
-                        }
-                        
-                        if(isset($mediaArray['is_image']) && $mediaArray['is_image'] == 1) {
+                            
+                            $conv_file_name = 'con_'.time().'.ogg';
+                            $cont_file_path = $this->config->item('post_media_path').$conv_file_name;
+                            shell_exec("/usr/local/bin/ffmpeg -i ".$this->config->item('post_media_path').$media_file." -acodec libvorbis -vcodec libtheora -f ogg ".$cont_file_path." 2>&1");
+                            if(file_exists($this->config->item('post_media_path').$media_file)) { unlink($this->config->item('post_media_path').$media_file); } 
+                            $media_file = $conv_file_name;
+                            
+                        } else if($file_type[0] == 'audio') {
+                            $media_type = 3; // for voice
+                            $conv_file_name = 'con_'.time().'.mp3';
+                            $cont_file_path = $this->config->item('post_media_path').$conv_file_name;
+                            shell_exec("/usr/local/bin/ffmpeg -i ".$this->config->item('post_media_path').$media_file." -f mp3 ".$cont_file_path." 2>&1");
+                            if(file_exists($this->config->item('post_media_path').$media_file)) { unlink($this->config->item('post_media_path').$media_file); } 
+                            $media_file = $conv_file_name;
+                            
+                        } else if($file_type[0] == 'image' && $mediaArray['is_image'] == 1) {
                             $media_type = 1;
                             // for resize the media image
                             $this->resize_image($media_file);
@@ -109,6 +113,19 @@ class Home extends CI_Controller {
                             if(file_exists($this->config->item('post_media_path').$media_file)) {
                                 unlink($this->config->item('post_media_path').$media_file);
                             }
+                        } else {
+                            if(file_exists($this->config->item('post_media_path').$media_file)) {
+                                unlink($this->config->item('post_media_path').$media_file); 
+                            } 
+                            $this->session->set_flashdata('class', 'alert-danger');
+                            $this->session->set_flashdata('post_message', 'File type is not allowed!');
+                            redirect('home');
+                        } 
+                        
+                        if(!file_exists($this->config->item('post_media_path').$media_file) && $mediaArray['is_image'] != 1) {
+                            $this->session->set_flashdata('class', 'alert-danger');
+                            $this->session->set_flashdata('post_message', 'Failed try again!');
+                            redirect('home');
                         }
                         
                     }
